Move pytest_configure into installed area.
This allows it to work when run via matplotlib.test().
