+++ lib/mpl_toolkits/tests/conftest.py
@@ -1,4 +1,5 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-from matplotlib.tests.conftest import mpl_test_settings
+from matplotlib.tests.conftest import (mpl_test_settings,
+                                       pytest_configure, pytest_unconfigure)
